rbi-ai-assistant/server/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
sentence-transformers
torch

pypdf
requests
httpx
